Create hook for pools & contribution
